requirements.txt
# --- API & config ---
fastapi>=0.114
uvicorn[standard]>=0.30
pydantic>=2.7
PyYAML>=6.0.1
python-dotenv==1.0.1
python-multipart>=0.0.9   # <— uploads multiples dans FastAPI
httpx>=0.27  
pymupdf>=1.24          # (utile pour tests d’intégration)

# --- LLM / embeddings ---
openai==1.79.0
tiktoken>=0.7              # (optionnel mais pratique)
google-generativeai>=0.5.2

# --- Extraction fichiers ---
pdfplumber>=0.11           # PDF texte (s’appuie sur pdfminer.six)
python-docx>=1.1
pandas>=2.2
openpyxl>=3.1
xlrd>=2.0

# --- OCR (optionnel, pour PDF scannés / images) ---
pillow>=10.4
pytesseract>=0.3.10      # nécessite Tesseract installé sur l’OS (+ langues fra/ara)

# --- Logs (si tu veux aller au-delà du logging stdlib) ---
structlog>=24.1

# --- Tests ---
pytest>=8.3
pytest-asyncio>=0.23

# --- Base de données ---
neo4j>=5.0

# --- langchain ---
langchain
langchain-openai
langchain-community
langchain-neo4j
langchain-huggingface
langchain-google-genai
google-genai

sentence-transformers
opencv-python-headless

fastmcp